feat: divide components and get data using graphql
